Move gesture detectors and listeners in the GestureHelper class/file
Revert automatic code cleanup
